back/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.1.0